Issue #154 refactor: api 패키지 리팩토링
