+++ app/src/androidTest/java/com/lateinit/rightweight/SharedRoutineTest.kt
@@ -3,7 +3,7 @@
 //import androidx.room.Room
 //import androidx.test.ext.junit.runners.AndroidJUnit4
 //import androidx.test.platform.app.InstrumentationRegistry
-//import com.lateinit.rightweight.data.RoutineApiService
+//import com.lateinit.rightweight.data.api.RoutineApiService
 //import com.lateinit.rightweight.data.database.AppDatabase
 //import com.lateinit.rightweight.data.database.mediator.*
 //import com.lateinit.rightweight.data.model.remote.FieldTransformsModelData
