--- app/src/main/java/com/lateinit/rightweight/data/database/baseEntity.kt
@@ -1,4 +1,0 @@
-package com.lateinit.rightweight.data.database
-
-class baseEntity {
-}
