+++ app/src/main/java/com/lateinit/rightweight/data/database/entity/HistoryExercise.kt
@@ -4,7 +4,7 @@ import androidx.room.ColumnInfo
 import androidx.room.Entity
 import androidx.room.ForeignKey
 import androidx.room.PrimaryKey
-import com.lateinit.rightweight.data.ExercisePartType
+import com.lateinit.rightweight.data.model.local.ExercisePartType
 
 @Entity(
     tableName = "history_exercise",
