+++ app/src/main/java/com/lateinit/rightweight/data/datasource/remote/LoginDataSource.kt
@@ -1,6 +1,6 @@
 package com.lateinit.rightweight.data.datasource.remote
 
-import com.lateinit.rightweight.data.LoginResponse
+import com.lateinit.rightweight.data.model.remote.LoginResponse
 
 interface LoginDataSource {
