+++ app/src/main/java/com/lateinit/rightweight/ui/login/LoginViewModel.kt
@@ -2,7 +2,7 @@ package com.lateinit.rightweight.ui.login
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.lateinit.rightweight.data.LoginResponse
+import com.lateinit.rightweight.data.model.remote.LoginResponse
 import com.lateinit.rightweight.data.model.local.User
 import com.lateinit.rightweight.data.repository.LoginRepository
 import com.lateinit.rightweight.data.repository.UserRepository
